refactor: split functional codes
