+++ src/App.js
@@ -1,17 +1,8 @@
-import ConvertInputTo from './modules/ConvertInputTo.js';
-import Print from './modules/Print.js';
-import Get from './modules/Get.js';
+import Fucntional from './functinoal/Functional.js';
 
 class App {
   async play() {
-    const lottoArray = await ConvertInputTo.lottoArray();
-    Print.purchasedLotto(lottoArray);
-    const lottoBoard = await ConvertInputTo.lottoBoard();
-    const lottoResult = Get.lottoResult(lottoArray, lottoBoard);
-    Print.lottoResult(lottoResult);
-    Print.lottoReturnRatio(
-      Get.lottoReturnRatio(lottoResult, lottoArray.length)
-    );
+    await Fucntional.play();
   }
 }
