+++ src/constants/STRINGS.js
@@ -2,8 +2,8 @@ const ERROR_HEADER = '[ERROR]';
 
 const STRINGS = Object.freeze({
   PURCHASE_PRICE_INPUT_MESSAGE: '구입금액을 입력해 주세요.\n',
-  WINNING_NUMBER_INPUT_MESSAGE: '\n당첨 번호를 입력해 주세요.\n',
-  BONUS_NUMBER_INPUT_MESSAGE: '\n보너스 번호를 입력해 주세요.\n',
+  WINNING_NUMBER_INPUT_MESSAGE: '당첨 번호를 입력해 주세요.\n',
+  BONUS_NUMBER_INPUT_MESSAGE: '보너스 번호를 입력해 주세요.\n',
 
   ERROR_HEADER,
