+++ src/functinoal/modules/ErrorCheck.js
@@ -1,6 +1,6 @@
 import Is from './Is.js';
 
-import CONSTANTS from '../constants/CONSTANTS.js';
+import CONSTANTS from '../../constants/CONSTANTS.js';
 
 const {
   NOT_AN_INTEGER_ERROR_MESSAGE,
